requirements.txt
asgiref==3.4.1
autopep8==1.5.7
check-python-versions==0.19.1
Django==3.0
django-filter==21.1
djangorestframework==3.10.0
djangorestframework-simplejwt==4.7.2
pycodestyle==2.7.0
PyJWT==2.2.0
pytz==2021.1
PyYAML==6.0
sqlparse==0.4.2
toml==0.10.2
